Move repository back to domain
