+++ src/domain/auth/entities/auth-payload.entity.ts
